# docs/openapi-v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ecom-ai API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/content/legal-documents/{document_type}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get Legal Document",
        "operationId": "LegalDocument_GetPublished",
        "parameters": [
          {
            "name": "document_type",
            "in": "path",
            "required": true,
            "schema": {
              "enum": [
                "terms_of_service",
                "privacy_policy"
              ],
              "type": "string",
              "title": "Document Type"
            }
          },
          {
            "name": "version",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 40,
              "title": "Version"
            }
          },
          {
            "name": "locale",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 16,
              "default": "zh-CN",
              "title": "Locale"
            }
          },
          {
            "name": "region_code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 16,
              "default": "CN",
              "title": "Region Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_LegalDocument_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Envelope_LegalDocument_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LegalDocument"
          },
          "meta": {
            "$ref": "#/components/schemas/ResponseMeta"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "Envelope[LegalDocument]"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "LegalDocument": {
        "properties": {
          "document_type": {
            "type": "string",
            "enum": [
              "terms_of_service",
              "privacy_policy"
            ],
            "title": "Document Type"
          },
          "document_version": {
            "type": "string",
            "title": "Document Version"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "locale": {
            "type": "string",
            "title": "Locale"
          },
          "region_code": {
            "type": "string",
            "title": "Region Code"
          },
          "safe_content": {
            "type": "string",
            "title": "Safe Content"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "effective_at": {
            "type": "string",
            "title": "Effective At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "document_type",
          "document_version",
          "title",
          "locale",
          "region_code",
          "safe_content",
          "content_hash",
          "effective_at"
        ],
        "title": "LegalDocument"
      },
...
      "ResponseMeta": {
        "properties": {
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          }
        },
        "type": "object",
        "title": "ResponseMeta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}